fix warning int <--> unisgned int
git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/trunk/rtt@6581 ce417995-dfc9-0310-95a0-acdaff106893

+++ tests/buffers_test.cpp
@@ -797,7 +797,7 @@ void BuffersTest::testAtomicQueue()
 
     // assert: sum queues == sum dequeues
     CPPUNIT_ASSERT( aworker->appends + bworker->appends + cworker->appends + grower->i 
-                    == aworker->erases + bworker->erases + cworker->erases + aqueue->size() );
+                    == aworker->erases + bworker->erases + cworker->erases + (int)(aqueue->size()) );
 
     delete aworker;
     delete bworker;
